Blocks Support
This commit adds support for message blocks.

Slack is moving towards this as the default method of posting
complex messages in future.

The blocks themselves are just another JSON object passed to the API, so
supporting them is simple.

This commit be @symroe was taken from
https://github.com/os/slacker/pull/152, but with a single commit
message. @genghis should get the credit for the initial work.